+++ CHANGELOG.rst
@@ -2,7 +2,13 @@
 Changelog
 =========
 
-0.0.1 (2016-9-21)
------------------------------------------
+0.0.1
+-----
 
 * First release on PyPI.
+
+0.0.2
+-----
+
+* The configfile command line argument added.
+* When no configfile argument present APEX will now search multiple default paths to find a configuration file.